+++ src/cluster/src/install.rs
@@ -15,15 +15,15 @@ use fluvio::config::{TlsPolicy, TlsConfig, TlsPaths, ConfigFile, Profile};
 use flv_util::cmd::CommandExt;
 use flv_future_aio::timer::sleep;
 use k8_client::{K8Client, ClientError as K8ClientError};
-use k8_config::{K8Config, KubeContext};
+use k8_config::{K8Config};
 use k8_client::metadata::MetadataClient;
 use k8_obj_core::service::ServiceSpec;
 use k8_obj_metadata::InputObjectMeta;
-use url::Url;
 use semver::Version;
 
 use crate::ClusterError;
 use crate::helm::{HelmClient, Chart, InstalledChart};
+use crate::check::get_cluster_server_host;
 
 const DEFAULT_NAMESPACE: &str = "default";
 const DEFAULT_REGISTRY: &str = "infinyon";
@@ -35,7 +35,6 @@ const DEFAULT_CHART_REMOTE: &str = "https://infinyon.github.io";
 const DEFAULT_GROUP_NAME: &str = "main";
 const DEFAULT_CLOUD_NAME: &str = "minikube";
 const DEFAULT_HELM_VERSION: &str = "3.2.0";
-const SYS_CHART_VERSION: &str = "0.2.0";
 
 /// Distinguishes between a Local and Remote helm chart
 #[derive(Debug)]
@@ -482,27 +481,8 @@ impl ClusterInstaller {
     /// Get installed system chart
     pub fn sys_charts() -> Result<Vec<InstalledChart>, ClusterError> {
         let helm_client = HelmClient::new()?;
-        let sys_charts = helm_client.installed_sys_charts(DEFAULT_CHART_SYS_REPO)?;
+        let sys_charts = helm_client.get_installed_chart_by_name(DEFAULT_CHART_SYS_REPO)?;
         Ok(sys_charts)
-    }
-
-    // Getting server hostname from K8 context
-    fn get_cluster_server_host(&self, kc_config: KubeContext) -> Result<String, IoError> {
-        if let Some(ctx) = kc_config.config.current_cluster() {
-            let server_url = ctx.cluster.server.to_owned();
-            let url = match Url::parse(&server_url) {
-                Ok(url) => url,
-                Err(e) => {
-                    return Err(IoError::new(
-                        ErrorKind::Other,
-                        format!("error parsing server url {}", e.to_string()),
-                    ))
-                }
-            };
-            Ok(url.host().unwrap().to_string())
-        } else {
-            Err(IoError::new(ErrorKind::Other, "no context found"))
-        }
     }
 
     /// Runs pre install checks
@@ -522,22 +502,12 @@ impl ClusterInstaller {
         // check installed system chart version
         let sys_charts = self
             .helm_client
-            .installed_sys_charts(DEFAULT_CHART_SYS_REPO)?;
-        if sys_charts.len() == 1 {
-            let installed_chart = sys_charts.first().unwrap();
-            let installed_chart_version = installed_chart.app_version.clone();
-            // checking version of chart found
-            if Version::parse(&installed_chart_version) < Version::parse(SYS_CHART_VERSION) {
-                return Err(ClusterError::Other(format!(
-                    "Fluvio system chart {} is not compatible with fluvio platform, please install version >= {}",
-                    installed_chart_version, SYS_CHART_VERSION
-                )));
-            }
-        } else if sys_charts.is_empty() {
+            .get_installed_chart_by_name(DEFAULT_CHART_SYS_REPO)?;
+        if sys_charts.is_empty() {
             return Err(ClusterError::Other(
                 "Fluvio system chart is not installed, please install fluvio-sys first".to_string(),
             ));
-        } else {
+        } else if sys_charts.len() > 1 {
             return Err(ClusterError::Other(
                 "Multiple fluvio system charts found".to_string(),
             ));
@@ -550,7 +520,7 @@ impl ClusterInstaller {
                 // ignore server check for pod
             }
             K8Config::KubeConfig(config) => {
-                let server_host = match self.get_cluster_server_host(config) {
+                let server_host = match get_cluster_server_host(config) {
                     Ok(server) => server,
                     Err(e) => {
                         return Err(ClusterError::Other(format!(
